% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSTZwritR.r
\name{ACTH7}
\alias{ACTH7}
\title{Dataset ACTH7.gpkg}
\description{
This data set is from Johnson et al. 2017, and slightly modified by us.
It is an eSTZ of Eriocoma thurberiana.
Johnson, R. C., E. A. Leger, and Ken Vance-Borland. "Genecology of Thurber’s Needlegrass (Achnatherum thurberianum [Piper] Barkworth) in the Western United States." Rangeland Ecology & Management 70.4 (2017): 509-517.
}
\section{ACTH7.gpkg}{


This data is, by default, used in orderZones()
}

