% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffR.R
\name{buffR}
\alias{buffR}
\title{buffR}
\usage{
buffR(x, buf_prcnt)
}
\arguments{
\item{x}{an input STZ, as vector or raster data -vector is much preferred for speed.}

\item{buf_prcnt}{the amount of buffering to add to the map, defaults to 0.025 or 2.5\%}
}
\value{
a bounding box around the eSTZ product of the specified buffer amount.
}
\description{
buffer an input STZ to determine map extents
}
\keyword{internal}
