% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSTZwritR.r
\docType{data}
\name{Carto_cities}
\alias{Carto_cities}
\title{select cities for cartography}
\format{
A spatial vector data set.
}
\usage{
Carto_cities
}
\description{
A geospatial vector data set of some US cities
Downloaded from https://simplemaps.com/data/us-cities and subset using R.
}
\details{
\itemize{
\City Full City name.
\State Abbreviated state name
}
}
\keyword{datasets}
\keyword{internal}
