% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSTZwritR.r
\docType{data}
\name{countries}
\alias{countries}
\title{Borders of three North American nations}
\format{
A data frame/tibble/sf with 3 rows and 2 columns:
\describe{
\item{name_long}{Country name}
\item{geom}{sf geometry column }
}
}
\usage{
countries
}
\description{
This data set contains the borders of three North American countries.
It is used within the \code{mapmaker} function to showcase the edges of both landmasses and where the
borders of the United States end. These data are subsetted from the 'spData' r package.
}
\keyword{datasets}
