% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirmakR.R
\name{dir_makr}
\alias{dir_makr}
\title{Create a directory and write out all empirical stz data}
\usage{
dir_makr(outpath, sci_name, nrcs_code, regioncode, overwrite, estz_type)
}
\arguments{
\item{outpath}{the directory folder where the final products will be placed. Note there is no need
to create an independent folder to hold them, this will be included in the product.}

\item{sci_name}{the scientific name of the taxon}

\item{nrcs_code}{from https://plants.usda.gov/}

\item{regioncode}{a character vector string of the name for this or, the output from \code{regioncoding}}

\item{overwrite}{whether to overwrite the existing directory (if it already exists)}

\item{estz_type}{one of 'lg' for landscape genetics (or genetic), 'cg' for common garden, or 'cm' for climate matched.
If multiple approaches choose the most robust method, e.g. 'cg' > 'lg' > 'cm'}

\item{x}{DATA PLACEHOLDER}
}
\description{
Create a directory and write out all empirical stz data
}
