% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSTZwritR.r
\docType{data}
\name{omernik}
\alias{omernik}
\title{Omernik level 3 ecoregions of the USA}
\format{
A data frame/tibble/sf with 3 rows and 2 columns:
\describe{
\item{US_L3CODE}{Ecoregion code}
\item{US_L3NAME}{Full ecoregion name}
\item{geometry}{sf geometry column}
}
}
\usage{
omernik
}
\description{
This data set contains the, slightly simplified, level 3 ecoregions for the USA.
These data are simplified to save space, and for use with cartographic, not for analytical purposes.
They are imported directly and silently by the \code{mapmaker} function to contextualize eSTZ border.
These data were sourced from the United States Geological Survey (usgs) \url{https://www.sciencebase.gov/catalog/item/55c77f7be4b08400b1fd82d4}
}
\keyword{datasets}
