% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderZones.R
\name{orderZones}
\alias{orderZones}
\title{Suggest seed zones names, or at least numeric orders, based on input polygons}
\usage{
orderZones(x, SeedZone, n, rasta, ...)
}
\arguments{
\item{x}{an sf/tibble/dataframe containing the spatial data which will be distributed as a final product}

\item{SeedZone}{name of the column containing the provisional SeedZones}

\item{n}{sample size of points across the entire shapefile for calculating an order, defaults to 2500. Note that this shouldn't
exceed the number of raster cells of the products used to generate the stz, or results will be replicated.}

\item{rasta}{Either a terra::spatrast object or, a character string of either "coarse" or 'cont' - short for contiguous USA.
GAI coarse has been aggregated by a factor of 2 from it's native 1km resolution to 4k, GAI cont is in the native resolution, but only covers the USA and a degree or so of Canada.
If neither of the rasters contained in the package meet your requirements, resubmit the GAI raster.}

\item{...}{additional parameters passed onto sf::st_sample}
}
\value{
A list containing three components, 'Reclassified' the shapefile which
was submitted, with the #' seedzone values over written by the ones resulting from the function.
'Summary' a dataframe containing the original zones, and final zones, as well as the calculated median and the number of samples used to calculate the median.
'Plot' a ggpubr boxplot of the results of a kruskal-wallis test amongst seed zones.
}
\description{
Given a shapefile of output seed transfer zones, with arbitrarily
assigned class or zone identifiers (e.g. numeric) this function specifies an
order to the zones from 1 to n zones reflecting an increase in humidity. In other
words, this function can assign an order to Seed Zones ID's, where drier zones
have lower numbers.

We have found it very difficult to discuss zones resulting from any PCA type
analysis, or incorporating measurements of phenology etc. Further it is very difficult
for us to estimate which zone a new population would be in, or notice many trends
in the zones at all. This function uses on aridity, as formulated in the Global
Aridity Index Version 3, as the ordering variable for zone numbering. For more
information on GAI please see https://www.nature.com/articles/s41597-022-01493-1.

Results may not always be statistically robust, but hopefully convey some
degree of useful information in their order regarding the range of the zones
across the species.
}
\examples{
acth7 <- sf::st_read(file.path(
  system.file(package="eSTZwritR"), "extdata", 'ACTH7.gpkg')
) |>
  sf::st_make_valid()

zoneOrder_suggestions <- orderZones(acth7, SeedZone = zone, n = 1000)

zoneOrder_suggestions$Summary
zoneOrder_suggestions$PlotKruskal
zoneOrder_suggestions$PlotDunns

acth7
zoneOrder_suggestions$Reclassified

# if you want to see the extent of the continental raster
# note that the Olympic peninsula and BC get so much rain they essentially
# swamps the colors of all other areas in plots.
cont <- terra::rast(
 file.path(
    system.file(package="eSTZwritR"), "extdata", 'GAI-cont.tif')
)
terra::plot(cont)
}
