% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Unified interior region boundaries}
\format{
A data frame/tibble/sf with 3 rows and 2 columns:
\describe{
\item{REG_NUM}{Region number}
\item{REG_NAME}{Region name}
\item{REG_ABB}{Unofficial 2-3-letter region abbreviation}
\item{geometry}{sf geometry column}
}
}
\usage{
regions
}
\description{
This data set contains the, slightly simplified, 12 department of interior regionas.
These data are simplified to save space, and for use with cartographic, not for analytical purposes.
They are imported directly and silently by the \code{regionCoding} function to determine which geographic areas to append to a filename.
These data are sourced from the US Department of Interior (DOI) \url{https://www.doi.gov/employees/reorg/unified-regional-boundaries}.
However, the 'REB_ABB' variable, which is utlimately used by eSTZwritR, is derived here and is not an officially sanctioned abbrevation by the USDOI.
}
\keyword{datasets}
