% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSTZwritR.r
\name{simplemaps_uscities_basicv1.79}
\alias{simplemaps_uscities_basicv1.79}
\title{Dataset Carto_cities}
\description{
This data set is curated to feature cities which can be helpful for contextualizing locations in a STZ map.
It is sourced from https://simplemaps.com/data/us-cities, and subst in a script available in the raw-data folder 'EssentialCities.R'.
}
\section{Carto_cities.gpkg}{


This data is used in mapmakR().
}

