% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderZones.R
\name{orderZones}
\alias{orderZones}
\title{Suggest seed zones names, or at least numeric orders, based on input polygons}
\usage{
orderZones(x, SeedZone, n, ...)
}
\arguments{
\item{x}{an sf/tibble/dataframe containing the spatial data which will be distributed as a final product}

\item{SeedZone}{name of the column containing the provisional SeedZones}

\item{n}{sample size of points across the entire shapefile for calculating an order, defaults to 2500. Note that this shouldn't
exceed the number of raster cells of the products used to generate the stz, or results will be replicated.}

\item{...}{additional parameters passed onto sf::st_sample}
}
\description{
This function seeks to make STZ more simple to convey, or at least visualize. Rather than using seed zone names
which relate to growth measurements, phenological responses, or combinations thereof, or zones from PCA axis we bastardize
the results of these analyses to two major components which humans think about extensively. Mean annual temperature, and
mean annual precipitation. By rescaling these two variables together we propose intuitively simple zones that follow a gradient from
hot and dry to cool and moist.
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package="sf")) |>
  dplyr::slice_head(n = 10)

pts <- orderZones(nc, SeedZone = 'NAME', n = 100)


SZplot <- ggplot2::ggplot() +
  ggplot2::geom_point(data = pts,
                     ggplot2::aes(
                       x = AnnualPrecipitation,
                       y = MeanAnnualTemp,
                       color = SeedZone)
 ) +
 ggplot2::labs(
   title = 'Relationship between Temperature and Precipitation across Seed Zones',
   color = 'Seed Zone',
   caption = 'Source: PRISM 1990-2020 climate normals') +
 ggplot2::theme_classic() +

 ggplot2::scale_y_continuous(
   'Mean Annual Temperature (C\u00B0)',
   ggplot2::sec.axis = ggplot2::sec_axis( ~ (. * 9/5) + 32, name = "MAT (F\u00B0)")
 ) +
 ggplot2::scale_x_continuous(
   'Annual Precipitation (mm)',
   ggplot2::sec.axis = ggplot2::sec_axis( ~ . * 0.0393701, name = "AP (inches)")
 )
}
