% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionCoding.R
\name{regionCoding}
\alias{regionCoding}
\title{Determine which interior regions the STZ should be marked with}
\usage{
regionCoding(x, n)
}
\arguments{
\item{x}{an empirical STZ as vector data.}

\item{n}{a sample size for determining which interior regions cover the most area of the stz
defaults to 1000, sizes above a couple thousand seem gratuitous.}
}
\value{
a list with a vector and a dataframe. The vector lists this component of the filename, at most
two interior regions separated by a '-'.
The dataframe contains a count of the number of randomly drawn points which intersect
interior regions. For areas with near ties we recommend increasing the sample size argument, \code{n} which is paseed to
to st:sample.
}
\description{
This function helps determine which Interior Regions coincide with the majority
of an empirical seed transfer zone and should be used for naming the file.
}
\examples{
acth7 <- sf::st_read(file.path(
  system.file(package="eSTZwritR"), "extdata", 'ACTH7.gpkg')
)

rc <- regionCoding(acth7)
rc$SuggestedName # name suggestions
rc$RegionsCovered # names suggested

}
